% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{assoc_plot}
\alias{assoc_plot}
\title{assoc_plot}
\usage{
assoc_plot(data, corr, ylab = NULL, title = NULL, subtitle = NULL,
  type = "log10p", x.min = NULL, x.max = NULL, legend=TRUE)
}
\arguments{
\item{data}{data.frame with markername (marker), chromosome (chr), position (pos) and either z-statistics (z) or probabilities (prob) columns}

\item{corr}{correlation matrix between markers}

\item{ylab}{the y-axis label}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{type}{the type of the plot either log10p or probabilities}

\item{x.min}{start of region}

\item{x.max}{end of region}
  
\item{legend}{add r2 legend}
}
\description{
assoc_plot plots a scatter graph of associations (e.g. log10 p-values)
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
