% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{stack_assoc_plot}
\alias{stack_assoc_plot}
\title{stack_assoc_plot}
\usage{
stack_assoc_plot(markers, z, corr, traits, x.min = NULL, x.max = NULL,
  top.marker = NULL, legend=TRUE)
}
\arguments{
\item{markers}{data.frame of markers with markername (marker), chromosome (chr) and position (pos)}

\item{z}{matrix of Z-scores with one column for each trait}

\item{corr}{matrix of correlation statistics between markers}

\item{traits}{trait names}

\item{x.min}{start of region}

\item{x.max}{end of region}

\item{top.marker}{the marker to be annotated in the plot}
  
\item{legend}{add r2 legend}
}
\description{
stack_assoc_plot plots stacked regional association plots
}
\author{
James R Staley <js16174@bristol.ac.uk>
}
